Feat [#5] CSV 파일에서 실수로 읽어짐에 따라 double로 수정

+++ src/main/java/org/drools/medical/Patient.java
@@ -13,12 +13,12 @@ import lombok.Setter;
 public class Patient {
 
 	private Long index;	// 입원 기록 ID
-	private int sbp;	// 수축기 혈압
-	private int respirationRate;	// 호흡률
+	private double sbp;	// 수축기 혈압
+	private double respirationRate;	// 호흡률
 	private Set<String> complaints;	// 불편 사항, 호소 증상
 
 	@Builder
-	public Patient(Long index, int sbp, int respirationRate, String complaints) {
+	public Patient(Long index, double sbp, double respirationRate, String complaints) {
 		this.index = index;
 		this.sbp = sbp;
 		this.respirationRate = respirationRate;
